[Austin] Restructure server files
