+++ package.json
@@ -10,11 +10,12 @@
     "pg": "^7.4.3"
   },
   "scripts": {
-    "start": "node index.js",
+    "start": "node src/server/index.js",
     "heroku-postbuild": "cd client && yarn && yarn run build",
-    "test": "NODE_ENV=test mocha ./routesTest --exit",
+    "test": "NODE_ENV=test mocha ./test --exit",
     "db:init": "scripts/initDb.js development test",
-    "db:clean": "scripts/cleanDb.js"
+    "db:clean": "scripts/cleanDb.js",
+    "client:build": "cd client && yarn build && rm -r ../src/server/static/ && mv build/ ../src/server/static/"
   },
   "devDependencies": {
     "chai": "^4.2.0",
