+++ scripts/cleanDb.js
@@ -1,8 +1,8 @@
 #!/usr/bin/env node
 
 // Clean up development database - tables
-const configFor = require('../db/config');
-const { deleteUser, deleteDb, deleteTables } = require('../db/manage');
+const configFor = require('../src/db/config');
+const { deleteUser, deleteDb, deleteTables } = require('../src/db/manage');
 
 function clean(config) {
   deleteDb(config.name);
