+++ src/server/routes/potentials.js
