+++ test/index.js
@@ -1,11 +1,11 @@
-const pool = require('../db');
-const { truncateTables } = require('../db/manage');
+const pool = require('../src/db');
+const { truncateTables } = require('../src/db/manage');
 
 beforeEach(async function resetDb() {
   await truncateTables(pool);
 });
 
-const app = require('../index')
+const app = require('../src/server/index')
 const TEST_PORT = 4000;
 const startApp = () => app.listen(TEST_PORT, () => {
     console.log(`Tests running on port ${TEST_PORT}`);
